Surround registry key containg the location of the exe with ""
If the location of "VirusTotal Uploader.exe" is not surrounded with
apostrophes in the registry key then Windows will not be able to run the
application when the context menu button is used if there is a space in
the path of the directory of the exe. This is because windows will
attempt to run everything before the space.

This commit adds apostrophes around the path in the registry key to
avoid this error.